requirements.txt
alembic==1.4.2
APScheduler==3.6.3
astropy==4.0
Babel==2.7.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Click==7.0
cryptography==3.0
cycler==0.10.0
docxtpl==0.6.3
Flask==1.1.1
Flask-Babel==0.12.2
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.2.0
lxml==4.5.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
paramiko==2.7.1
patsy==0.5.1
pyarrow==0.15.1
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.7
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.22.0
scipy==1.4.1
six==1.15.0
soupsieve==2.0
SQLAlchemy==1.3.12
statsmodels==0.11.1
tzlocal==2.1
unoconv==0.9.0
urllib3==1.25.10
Werkzeug==0.16.0
windrose==1.6.7
XlsxWriter==1.3.1
